bs4 task changed

+++ bs4_examples/main.py
@@ -8,15 +8,21 @@ from scraping_1 import get_file
 
 FILE_PATH = 'html_content.html'
 DATA_PATH = 'books.csv'
+DATA_PATH_JSON = 'books.json'
 
 
-def main(file_path, data_path):
+def main(file_path, data_path, data_path_json):
     if not os.path.exists(file_path):
         get_file(file_path)
 
-    if not os.path.exists(data_path):
+    if not os.path.exists(data_path) or not os.path.exists(data_path_json):
         df = get_df(file_path)
-        df.to_csv(data_path, index=False)
+        if not os.path.exists(data_path):
+            df.to_csv(data_path, index=False)
+        if not os.path.exists(data_path_json):
+            json_data = df.to_json(orient='records', indent=4)
+            with open('books.json', 'w', encoding='utf-8', ) as f:
+                f.write(json_data)
 
     df = pd.read_csv(data_path)
     index = random.randint(1, 1000)
@@ -28,4 +34,4 @@ def main(file_path, data_path):
 
 
 if __name__ == '__main__':
-    main(FILE_PATH, DATA_PATH)
+    main(FILE_PATH, DATA_PATH, DATA_PATH_JSON)
